fix(security): use proper uuid for device token
Okta device tokens are supposed to be unguessable. They are meant to
they help in not having to MFA constantly for devices, essentially a
short-cirtuit to avoid needing MFA. So in sensitivity they are between
your password and your MFA, therefore they are NOT meant to be
precomputed from the username (as saml2aws does it today).

If the device-token is easily guessable, then it defeats MFA, and users
The current users of saml2aws live in a false sense of security when
they allow remember-my-device (the default with saml2aws).

With saml2aws using a predictable device token, it poses a security risk
for any user using saml2aws, as it defeats using MFA.

This fix aims at generating a UUID specific to the device saml2aws runs
on and storing it as a secret (like storing the users password).
This makes using 'remember device' safe.